refactor: 컨벤션
